--- app/src/main/java/com/otiutiunnyk/diploma/parking_app/ui/screen/ExploreScreen.kt
@@ -1,9 +1,0 @@
-package com.otiutiunnyk.diploma.parking_app.ui.screen
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ExploreScreen() {
-    Text(text = "Explore Screen") //TODO: default screen
-}
